title: Title_DemoStreamTask without matching hacks
The goto ladder and register pins stood for a plain switch whose cases each
advance the state and break; jump2 merges the repeated tails. Stream_RestoreAfterLoad
returns u16, which every caller's mask already implied.

+++ include/main/stream.h
@@ -37,7 +37,7 @@ void        Mdec_SetupBuffers(u8* arg0);
 void        Mdec_ResolveStreamBuffer(u8* arg0);
 void        Stream_InitFromSlot(u32 arg0);
 u32         func_8001F180(u32 arg0);
-s32         Stream_RestoreAfterLoad(s32 arg0, s32 arg1);
+u16         Stream_RestoreAfterLoad(s32 arg0, s32 arg1);
 void        func_8001F430(void);
 void        Mdec_UploadSlice(void);
 void        Mdec_KickStrip(void);
